Hover functionality to view car image in the admin dashboard, added the owner id linking to the add car form

+++ PHP/admin_dashboard/car_list.php
@@ -1,3 +1,10 @@
+<style>
+    .car-image {
+        cursor: pointer;
+        width: 100px;
+    }
+</style>
+
 <section>
     <h2>List of Cars</h2>
     <table>
@@ -10,6 +17,7 @@
                 <th>Kilometers</th>
                 <th>Owner</th>
                 <th>Price</th>
+                <th>Available</th>
                 <th>Actions</th>
             </tr>
         </thead>
@@ -20,27 +28,33 @@
             $cars = mysqli_query($conn, $sql);
             foreach ($cars as $car) {
                 // Display car information and actions
-                echo "<tr onclick=\"toggleCarImage({$car['id']})\">"; // Call JavaScript function when clicking on a row
-                echo "<td>{$car['id']}</td>";
-                echo "<td>{$car['brand']}</td>";
-                echo "<td>{$car['model']}</td>";
-                echo "<td>{$car['color']}</td>";
+                echo "<tr>"; // Call JavaScript function when clicking on a row
+                echo "<td class='car-image' data-id='{$car['id']}'>{$car['id']}</td>";
+                echo "<td class='car-image' data-id='{$car['id']}'>{$car['brand']}</td>";
+                echo "<td class='car-image' data-id='{$car['id']}'>{$car['model']}</td>";
+                echo "<td class='car-image' data-id='{$car['id']}'>{$car['color']}</td>";
                 echo "<td>{$car['km']}</td>";
-                echo "<td>{$car['owner_id']}</td>";
+                echo "<td class='user-info' data-id='{$car['owner_id']}'>{$car['owner_id']}</td>";
                 echo "<td>{$car['price']}</td>";
-                echo "<td><a href=\"http://localhost/Mini-PHP-Project/PHP/delete_car.php?id={$car['id']}\">Delete</a>";
-                // Add update option if the user is the owner of the car
-                if ($car['owner_id'] == $_SESSION['user_id']) {
-                    echo " | <a href=\"../update_car.php?id={$car['id']}\">Update</a>";
+                $current_date = date('Y-m-d');
+                $car_commanded_sql = "SELECT * FROM command WHERE car_id={$car['id']} AND start_date <= '$current_date' AND end_date >= '$current_date'";
+                $car_commanded = mysqli_query($conn, $car_commanded_sql);
+                $available = mysqli_num_rows($car_commanded) == 0;
+                if (!$available) {
+                    echo "<td>Not Available</td>";
+                } else {
+                    echo "<td>Available</td>";
                 }
-                echo "</td>";
-                echo "</tr>";
-                // Display car image row
-                echo "<tr class=\"car-image-row\">";
-                echo "<td colspan=\"8\">";
-                echo "<img id=\"car-image-{$car['id']}\" class=\"car-image\" src=\"http://localhost/Mini-PHP-Project/Resources/Images/car_images/{$car['image']}\" alt=\"Car Image\">"; // Image element with unique ID
-                echo "</td>";
-                echo "</tr>";
+                if ($available) {
+                    echo "<td><a href=\"http://localhost/Mini-PHP-Project/PHP/delete_car.php?id={$car['id']}\">Delete</a>";
+                    // Add update option if the user is the owner of the car
+                    if ($car['owner_id'] == $_SESSION['user_id']) {
+                        echo " | <a href=\"../update_car.php?id={$car['id']}\">Update</a>";
+                    }
+                }
+                else {
+                    echo "<td>Car is in use</td>";
+                }
             }
             ?>
         </tbody>
